Extract the get script logic as a single utility file.
