+++ src/special/bin.js
@@ -1,26 +1,5 @@
 import path from 'path';
-import yaml from 'js-yaml';
-
-const travisCommands = [
-  // Reference: http://docs.travis-ci.com/user/customizing-the-build/#The-Build-Lifecycle
-  'before_install',
-  'install',
-  'before_script',
-  'script',
-  'after_success or after_failure',
-  'before_deploy',
-  'deploy',
-  'after_deploy',
-  'after_script',
-];
-
-function concat(array, item) {
-  return array.concat(item);
-}
-
-function getObjectValues(object) {
-  return Object.keys(object).map(key => object[key]);
-}
+import getScripts from '../utils/get-scripts';
 
 function toKeyValuePair(object) {
   return Object.keys(object).map(key => ({ key, value: object[key] }));
@@ -72,14 +51,10 @@ function getUsedDeps(deps, scripts, dir) {
   .map(metadata => metadata.dep);
 }
 
-export default (content, filename, deps, dir) => {
-  const basename = path.basename(filename);
-  if (basename === 'package.json') {
-    const scripts = getObjectValues(JSON.parse(content).scripts || {});
-    return getUsedDeps(deps, scripts, dir);
-  } else if (basename === '.travis.yml') {
-    const metadata = yaml.safeLoad(content) || {};
-    const scripts = travisCommands.map(cmd => metadata[cmd] || []).reduce(concat, []);
+export default (content, filepath, deps, dir) => {
+  const basename = path.basename(filepath);
+  if (basename === 'package.json' || basename === '.travis.yml') {
+    const scripts = getScripts(filepath, content);
     return getUsedDeps(deps, scripts, dir);
   }
